Anime API (Turing Anime Club App)
The Genres Dataset is the same as last time. But I have found an API that will give us access to 1500+ Anime Titles
My Pitch:
As head of the #Anime group at Turing, it is challenging for us to conduct weekly polls and see the engagement that we would like.
Creating an app as a tool to use to help manage polls, to - watchlist, completed - watchlist, and discussions would be a great legacy to leave behind when I can no longer lead.